// src/app/models/common/dataview-list-model.ts
export class DataviewListModel {
    
    id: number=0;

    imagePath: string='';    
    
    image: boolean=false;    

    name: string='';

    mainDescription: string='';

    secundaryDescription: string='';

    mainDescriptionSide: string='';

    secundaryDescriptionSide: string='';

    indPerson?:number

    
    // imagePathEmpaque: string ="assets/layout/images/empaque.png";

    // imagePathAvatar: string ="assets/layout/images/avatar-unisex.jpg";

}